WIP: Clean up translations
This is the first of many commits that will attempt to clean up the
translations in the site. The plan is to finish cleaning up the html and
replace text with their respective variables, moving the text to
translation files. Once that's complete, I'll start going through
translations and putting their text in translation files. Once all
translations are "processed", I'll delete all the translation copies and
re-structure the routing so that everything loads from one set of code
but renders the translations via the translation files. -- I realize
that may sound confusing.. More to come!