Hiding cryptobroker as requested in #9 (#10)
Cryptobroker has been **very** delayed in transferring coins to their
customers. This has led to a lot of issues with people being worried
that they may not get paid. For the time being, we're not going to list
cryptobroker on the site to ensure new members don't get subjected to
that sort of uncertainty.

This closes #9.